正在写菜品批量删除与修改菜品信息接口

+++ sky-server/src/main/java/com/sky/service/impl/DishServiceImpl.java
@@ -2,12 +2,16 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.dto.DishDTO;
 import com.sky.dto.DishPageQueryDTO;
 import com.sky.entity.Dish;
+import com.sky.exception.DeletionNotAllowedException;
+import com.sky.exception.SetmealEnableFailedException;
 import com.sky.mapper.DishFlavorMapper;
 import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.DishService;
 import com.sky.vo.DishVO;
@@ -26,6 +30,8 @@ public class DishServiceImpl implements DishService {
     private DishMapper dishMapper;
     @Resource
     private DishFlavorMapper dishFlavorMapper;
+    @Resource
+    private SetmealMapper setmealMapper;
 
     @Override
     public void addDish(DishDTO dishDTO) {
@@ -55,9 +61,34 @@ public class DishServiceImpl implements DishService {
 
     @Override
     public void deleteBatch(List<Long> dishIds) {
+        // 1.判断是否有套餐包含了该菜品
+        if(setmealMapper.findDishByDishId(dishIds)){
+            throw new SetmealEnableFailedException(MessageConstant.DISH_BE_RELATED_BY_SETMEAL);
+            // 2.判断菜品状态是否是在售状态
+        }else if (dishMapper.findDishStatus(dishIds)){
+            throw new DeletionNotAllowedException(MessageConstant.DISH_ON_SALE);
+            // 3.判断是否删除成功
+        }else if (dishMapper.deleteBatch(dishIds)){
+            dishFlavorMapper.deleteBatch(dishIds);
+        }
+    }
 
-        dishMapper.deleteBatch(dishIds);
-        dishFlavorMapper.deleteBatch(dishIds);
+    @Override
+    public void updateDish(DishDTO dishDTO) {
+        Dish dish = Dish.builder().build();
+        BeanUtils.copyProperties(dishDTO,dish);
+        if(dishMapper.updateDish(dish)){
+            dishFlavorMapper.updateDishFlavor(dishDTO.getFlavors(),dishDTO.getId());
+        }
+    }
+
+    @Override
+    public DishVO findById(Long dishId) {
+        DishVO dishVO = dishMapper.findById(dishId);
+        if(dishVO != null){
+            dishVO.setFlavors(dishFlavorMapper.findFlavorByDishId(dishId));
+        }
+        return dishVO;
     }
 
 }
